# internal/models/model.go
package models

// APIResponse defines a standard response structure
type APIResponse struct {
	Status     string      `json:"status"`
	StatusCode int         `json:"status_code"`
	Data       interface{} `json:"data"`
	Message    string      `json:"message"`
}